+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - linear source inversion *lsrcinv* in *nessi.signal*
 - *normalize* method for SUdata()
+- *nessidata* the NESSI data structure 
+- *suformat* a new SU format support compatible with the NESSI data structure
 
 ### Modified
 - *pfilter* method for SUdata()